fixe typo

+++ config/mainColoredBitmap.c
@@ -8,8 +8,8 @@
 
 int testMethod(int input1, int input2, char* unit, float output){
     char message[100];
-    int result = coloredBitmap(input1, input2, unit);
-    sprintf(message, "Inputs (%d, %d, %s). Expected output should be %f. Actuel output = %d", input1, input2, unit, output, result);
+    float result = coloredBitmap(input1, input2, unit);
+    sprintf(message, "Inputs (%d, %d, %s). Expected output should be %f. Actuel output = %f", input1, input2, unit, output, result);
     assert_with_message(((result >= output - 0.2) && (result <= output + 0.2)), message);
 }
